Moved heartbeat service
