+++ WebApp/backend/src/devices/devices.module.ts
@@ -4,6 +4,7 @@ import { DevicesConfigurationController } from './configuration/configuration.co
 import { ConfigurationService } from './configuration/configuration.service';
 import { DevicesController } from './devices.controller';
 import { HeartbeatController } from './heartbeat/heartbeat.controller';
+import { HeartbeatService } from './heartbeat/heartbeat.service';
 import { LogsController } from './logs/logs.controller';
 import { RestartController } from './restart/restart.controller';
 import { StatesController } from './states/states.controller';
@@ -11,7 +12,7 @@ import { StatusController } from './status/status.controller';
 
 @Module({
   imports: [DatabaseModule],
-  providers: [ConfigurationService],
+  providers: [ConfigurationService, HeartbeatService],
   controllers: [
     DevicesController,
     DevicesConfigurationController,
