+++ WebApp/backend/src/devices/heartbeat/heartbeat.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get, Param, Query, UseGuards } from '@nestjs/common';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
-import { HeartbeatService } from 'src/database/heartbeat/heartbeat.service';
+import { HeartbeatService } from 'src/devices/heartbeat/heartbeat.service';
 
 @Controller(['devices/:deviceId/heartbeat', 'devices/:deviceId/ticks'])
 export class HeartbeatController {
